Remove command line interface
Remvoe index.ts, which used to provide bin/cli features that called
wsdl-to-flow.

I have no need for that and will be using this only through the
wsdl-to-flow.ts API anyway, so fixing issues in the index.ts
implementation and adding tests for it isn't worth it for me right now.

CLI usage option can be added back later if there's demand for it.